requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
httpx
supabase-py
passlib[bcrypt]
python-jose[cryptography]
transformers
torch
scikit-learn
pandas
dp-agent
deepdiff